added shell of windows service and prepared for setuptools
