+++ setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup, find_packages
+setup(
+    name = "pyQBWC",
+    version = "0.1",
+    packages = find_packages(),
+)
